typehint and utils update, csv_service rework
