+++ src/utils/date_utlis.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta, date
+from typing import Generator
 
 
 def get_previous_date() -> date:
@@ -9,6 +10,6 @@ def get_current_date() -> date:
     return datetime.today().date()
 
 
-def get_dates(start_date=date(2003, 1, 1), end_date=date.today()) -> date:
+def get_dates(start_date=date(2003, 1, 1), end_date=date.today()) -> Generator[date, None, None]:
     for n in range(int((end_date - start_date).days)):
         yield start_date + timedelta(n)
