+++ src/weather_collector/scrapper/weather_scrapper.py
@@ -1,4 +1,5 @@
 from datetime import date
+from typing import Generator
 
 from src.weather_collector.crawler.weather_crawler import WeatherCrawler
 from src.weather_collector.parser.weather_parser import WeatherParser
@@ -10,7 +11,7 @@ class WeatherScrapper(AbstractScrapper):
         self.crawler = WeatherCrawler()
         self.parser = WeatherParser()
 
-    async def scrap_content(self, date_list: list[date]) -> list[dict]:
+    async def scrap_content(self, date_list: Generator[date, None, None]) -> list[dict]:
         result = []
         async for html, url, current_date in self.crawler.crawl_content(date_list):
             data = await self.parser.parse_content(html, url, current_date)
